+++ plugins/tasks/resource/src/main/java/com/walmartlabs/concord/plugins/resource/ResourceTask.java
@@ -21,11 +21,16 @@ package com.walmartlabs.concord.plugins.resource;
  */
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.walmartlabs.concord.sdk.Constants;
+import com.walmartlabs.concord.sdk.Context;
 import com.walmartlabs.concord.sdk.InjectVariable;
 import com.walmartlabs.concord.sdk.Task;
 
 import javax.inject.Named;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -42,8 +47,36 @@ public class ResourceTask implements Task {
 
     @SuppressWarnings("unused")
     public Object asJson(String path) throws IOException {
-        byte[] ab = Files.readAllBytes(Paths.get(path));
-        return new ObjectMapper().readValue(ab, Object.class);
+        return asJson(null, path, false);
+    }
+
+    @SuppressWarnings("unused")
+    public Object asJson(@InjectVariable("context") Context ctx, String path, boolean eval) throws IOException {
+        try (InputStream in = Files.newInputStream(Paths.get(path))) {
+            Object result = new ObjectMapper().readValue(in, Object.class);
+            if (eval) {
+                return ctx.interpolate(result);
+            } else {
+                return result;
+            }
+        }
+    }
+
+    @SuppressWarnings("unused")
+    public Object asYaml(String path) throws IOException {
+        return asYaml(null, path, false);
+    }
+
+    @SuppressWarnings("unused")
+    public Object asYaml(@InjectVariable("context") Context ctx, String path, boolean eval) throws IOException {
+        try (InputStream in = Files.newInputStream(Paths.get(path))) {
+            Object result = new ObjectMapper(new YAMLFactory()).readValue(in, Object.class);
+            if (eval) {
+                return ctx.interpolate(result);
+            } else {
+                return result;
+            }
+        }
     }
 
     @SuppressWarnings("unused")
@@ -51,10 +84,12 @@ public class ResourceTask implements Task {
         Path baseDir = Paths.get(workDir);
         Path tempDir = assertTempDir(baseDir);
 
-        Path resourceFile = Files.createTempFile(tempDir, "resource_", ".json");
-        new ObjectMapper().writerWithDefaultPrettyPrinter().writeValue(resourceFile.toFile(), content);
+        Path tmpFile = Files.createTempFile(tempDir, "resource_", ".json");
+        try (OutputStream out = Files.newOutputStream(tmpFile)) {
+            new ObjectMapper().writerWithDefaultPrettyPrinter().writeValue(out, content);
+        }
 
-        return baseDir.relativize(resourceFile.toAbsolutePath()).toString();
+        return baseDir.relativize(tmpFile.toAbsolutePath()).toString();
     }
 
     @SuppressWarnings("unused")
@@ -69,7 +104,7 @@ public class ResourceTask implements Task {
     }
 
     private Path assertTempDir(Path baseDir) throws IOException {
-        Path p = baseDir.resolve(".tmp");
+        Path p = baseDir.resolve(Constants.Files.CONCORD_TMP_DIR_NAME);
         if (!p.toFile().exists()) {
             Files.createDirectories(p);
         }