+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- resource-tasks: new method to read YAML files. Optional support for
+string interpolation in JSON and YAML files;
 - ansible: initial support for external roles.
 
 ### Changed